+++ components/ir/cartes-impact/gagnants-perdants/GagnantsPerdantsContent/index.ts
