+++ components/ir/cartes-impact/gagnants-perdants/gagnants-perdants-component.tsx
@@ -14,10 +14,10 @@ import React, { Fragment, PureComponent } from "react";
 // eslint-disable-next-line no-unused-vars
 import { connect, ConnectedProps } from "react-redux";
 
-import { fetchSimPop, simulateCasTypes } from "../../../redux/actions";
+import { fetchSimPop, simulateCasTypes } from "../../../../redux/actions";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../../redux/reducers";
-import { Card } from "../../common";
+import { RootState } from "../../../../redux/reducers";
+import { Card } from "../../../common";
 import styles2 from "./gagnants-perdants-component.module.scss";
 import { GagnantsPerdantsContent } from "./GagnantsPerdantsContent";
