+++ components/ir/cartes-impact/gagnants-perdants/index.ts
