+++ components/ir/cartes-impact/impact-component.jsx
@@ -2,8 +2,8 @@ import Grid from "@material-ui/core/Grid";
 import PropTypes from "prop-types";
 import { PureComponent } from "react";
 
-import CarteEtat from "../carte-etat";
-import InformationPanel from "../information-panel";
+import CarteEtat from "../../carte-etat";
+import InformationPanel from "../../information-panel";
 import { GagnantsPerdantsCard } from "./gagnants-perdants";
 import styles from "./impact-component.module.scss";
 import SimpleCard from "./simple-card";
