+++ components/ir/cartes-impact/impact-component.module.scss
