+++ components/ir/cartes-impact/index.ts
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../redux/reducers";
+import { RootState } from "../../../redux/reducers";
 import ImpactComponent from "./impact-component";
 
 const mapStateToProps = ({
