+++ components/ir/cartes-impact/simple-card/index.ts
@@ -1,8 +1,8 @@
 import { connect } from "react-redux";
 
-import { removeCasType, showEditCasTypesPopin, simulateCasTypes } from "../../../redux/actions";
+import { removeCasType, showEditCasTypesPopin, simulateCasTypes } from "../../../../redux/actions";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../../redux/reducers";
+import { RootState } from "../../../../redux/reducers";
 import SimpleCard from "./simple-card-component";
 
 const mapStateToProps = ({ descriptions, results }: RootState, { index }) => {
