+++ components/ir/cartes-impact/simple-card/simple-card-component.jsx
@@ -16,10 +16,10 @@ import { get } from "lodash";
 import PropTypes from "prop-types";
 import React, { Fragment } from "react";
 
-import { Card } from "../../common";
-import DoublePalmTreeIcon from "../../icons/double-palm-tree";
-import { NeutralTooltip } from "../../common";
-import { formatNumber } from "../../utils";
+import { Card } from "../../../common";
+import DoublePalmTreeIcon from "../../../icons/double-palm-tree";
+import { NeutralTooltip } from "../../../common";
+import { formatNumber } from "../../../utils";
 import SimpleCardImpactImpots from "./impact-impots";
 
 const RESIDENCE_ITEMS = [
