+++ components/ir/index.ts
@@ -1,1 +1,2 @@
 export { default as Articles } from "./articles";
+export { default as CartesImpact } from "./cartes-impact";
